third_party/WebKit/LayoutTests/animations/usecounters/playbackRate-usecounter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body></body>
<script>
'use strict';
// From UseCounter.h
var AnimationSetPlaybackRateCompensatorySeek = 2162;
var MS_PER_SEC = 1000;
function createDiv(test) {
  var element = document.createElement('div');
  document.body.appendChild(element);
  return element;
}

test(() => {
  var div = createDiv();
  var animation = div.animate(null, 100 * MS_PER_SEC);
  var startTimeBefore = animation.startTime;
  animation.playbackRate = 0.5;

  assert_equals(startTimeBefore, null);
  assert_false(internals.isUseCounted(document, AnimationSetPlaybackRateCompensatorySeek));
  animation.playbackRate = 2;
  assert_equals(animation.startTime, null);
  assert_false(internals.isUseCounted(document, AnimationSetPlaybackRateCompensatorySeek));
}, 'Check to ensure that we don\'t UseCount when the animation hasn\'t started and you set playbackRate');

test(() => {
  var div = createDiv();
  var animation = div.animate(null, 100 * MS_PER_SEC);
  return animation.ready.then(function() {
    var startTimeBefore = animation.startTime;
    animation.currentTime = 7 * MS_PER_SEC; // ms
    animation.playbackRate = 0.5;

    assert_equals(startTimeBefore, animation.startTime);
    assert_false(internals.isUseCounted(document, AnimationSetPlaybackRateCompensatorySeek));
    animation.playbackRate = 0.5;
    assert_equals(startTimeBefore, animation.startTime);
    assert_false(internals.isUseCounted(document, AnimationSetPlaybackRateCompensatorySeek));
  })
}, 'Check to ensure that we don\'t UseCount when you don\'t cause a compensatory seek by setting playbackRate');

test(() => {
  var div = createDiv();
  var animation = div.animate(null, 100 * MS_PER_SEC);
  return animation.ready.then(function() {
    var startTimeBefore = animation.startTime;
    animation.currentTime = 7 * MS_PER_SEC; // ms
    animation.playbackRate = 0.5;

    assert_equals(startTimeBefore, animation.startTime);
    assert_false(internals.isUseCounted(document, AnimationSetPlaybackRateCompensatorySeek));
    animation.playbackRate = 2;
    assert_not_equals(startTimeBefore, animation.startTime);
    assert_true(internals.isUseCounted(document, AnimationSetPlaybackRateCompensatorySeek));
  })
}, 'Check to ensure that we UseCount when you cause a compensatory seek by setting playbackRate');
</script>
